veyron_agency/requirements.txt
agency-swarm>=0.2.0
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0 